examples/projects/web-scraping/requirements.txt
newspaper4k==0.9.3.1
news-please==1.6.14
